work around nvJitLink symbol issues, fix a few more test skips, other misc. fixes

+++ python/pylibwholegraph/pylibwholegraph/tests/wholegraph_torch/ops/test_graph_add_csr_self_loop.py
@@ -37,7 +37,7 @@ def routine_func(**kwargs):
         target_node_count,
         edge_num,
         neighbor_node_count=neighbor_node_count,
-        csr_row_dtype=torch.int64,
+        csr_row_dtype=torch.int32,
         csr_col_dtype=torch.int32,
         weight_dtype=torch.float32,
     )
